services/integration-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic-settings==2.8.1
PyJWT==2.10.1
qdrant-client==1.14.2
python-dotenv==1.0.1

pytest==8.3.4
httpx==0.28.1
